assignment-4/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.0.1
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-cloud==0.34.0
google-genai==1.3.0
gotrue==2.11.4
graphviz==0.20.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.9.1
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.1.0
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==2.2.3
openai==0.28.0
openpyxl==3.1.5
package_name==0.1
packaging==24.2
pandas==2.2.3
parso==0.8.3
pillow==11.1.0
platformdirs==4.2.0
postgrest==0.19.3
prompt-toolkit==3.0.43
propcache==0.3.0
protobuf==5.29.3
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.1
pywin32==306
pyzmq==25.1.2
realtime==2.4.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
seaborn==0.13.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
storage3==0.11.3
streamlit==1.42.1
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
toml==0.10.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
